TWILIO/requirements.txt
 fastapi 
 uvicorn 
 twilio 
 python-multipart 
 jinja2
 python-dotenv